Cart store manager policy 1.5 1.6 2.2 tests part4 6356

+++ SadnaSrc/SadnaSrc/UserSpot/UserPolicyService.cs
@@ -25,6 +25,10 @@ namespace SadnaSrc.UserSpot
         }       
         public static void PromoteStorePolicies(string userName,string store,StoreManagerPolicy.StoreAction[] actionsToAdd)
         {
+            if (!_userDL.IsUserNameExist(userName))
+            {
+                throw new UserException(PromoteStoreManager.NoUserFound, "No user by the name '" + userName + " has been found for promotion!");
+            }
             foreach (StoreManagerPolicy.StoreAction oldAction in Enum.GetValues(typeof(StoreManagerPolicy.StoreAction)))
             {
                 _userDL.DeleteUserStorePolicy(userName, new StoreManagerPolicy(store, oldAction));
@@ -56,8 +60,12 @@ namespace SadnaSrc.UserSpot
                 if(policy.Store.Equals(store))
                 storePolicies.Add(policy);
             }
+            return SortStorePolicy(storePolicies.ToArray());
+        }
 
-            return storePolicies.ToArray();
+        private static StoreManagerPolicy[] SortStorePolicy(StoreManagerPolicy[] storePolicies)
+        {
+            return storePolicies.OrderBy(x => x.Action).ToArray();
         }
 
         public void AddStoreOwnership(string store)
